fix(test): dedupe otel schema DDL
